+++ src/renderer/src/modules/entry-column/components/mark-all-button.tsx
