+++ src/renderer/src/modules/entry-column/item.tsx
@@ -1,36 +1,27 @@
-import { useGeneralSettingKey } from "@renderer/atoms/settings/general"
-import { ListItemHoverOverlay } from "@renderer/components/ui/list-item-hover-overlay"
 import { LoadingCircle } from "@renderer/components/ui/loading"
 import { views } from "@renderer/constants"
-import { useAsRead } from "@renderer/hooks/biz/useAsRead"
-import { useEntryActions } from "@renderer/hooks/biz/useEntryActions"
-import { useFeedActions } from "@renderer/hooks/biz/useFeedActions"
-import { useNavigateEntry } from "@renderer/hooks/biz/useNavigateEntry"
-import { useRouteParamsSelector } from "@renderer/hooks/biz/useRouteParams"
 import { useAuthQuery } from "@renderer/hooks/common"
 import { FeedViewType } from "@renderer/lib/enum"
-import { showNativeMenu } from "@renderer/lib/native-menu"
 import { cn } from "@renderer/lib/utils"
 import { Queries } from "@renderer/queries"
 import type { FlatEntryModel } from "@renderer/store/entry"
-import { entryActions } from "@renderer/store/entry"
 import { useEntry } from "@renderer/store/entry/hooks"
 import type { FC, ReactNode } from "react"
-import { memo, useCallback } from "react"
-import { useDebounceCallback } from "usehooks-ts"
+import { memo } from "react"
 
 import { ReactVirtuosoItemPlaceholder } from "../../components/ui/placeholder"
-import { ArticleItem, ArticleItemSkeleton } from "./article-item"
-import { AudioItem, AudioItemSkeleton } from "./audio-item"
+import { ArticleItem, ArticleItemSkeleton } from "./Items/article-item"
+import { AudioItem, AudioItemSkeleton } from "./Items/audio-item"
 import {
   NotificationItem,
   NotificationItemSkeleton,
-} from "./notification-item"
-import { PictureItem, PictureItemSkeleton } from "./picture-item"
-import { SocialMediaItem, SocialMediaItemSkeleton } from "./social-media-item"
+} from "./Items/notification-item"
+import { PictureItem, PictureItemSkeleton } from "./Items/picture-item"
+import { SocialMediaItem, SocialMediaItemSkeleton } from "./Items/social-media-item"
+import { VideoItem, VideoItemSkeleton } from "./Items/video-item"
+import { EntryItemWrapper } from "./layouts/EntryItemWrapper"
 import { girdClassNames } from "./styles"
 import type { EntryListItemFC } from "./types"
-import { VideoItem, VideoItemSkeleton } from "./video-item"
 
 interface EntryItemProps {
   entryId: string
@@ -43,17 +34,6 @@ function EntryItemImpl({
   entry: FlatEntryModel
   view?: number
 }) {
-  const { items } = useEntryActions({
-    view,
-    entry,
-  })
-
-  const { items: feedItems } = useFeedActions({
-    feedId: entry.feedId,
-    view,
-    type: "entryList",
-  })
-
   const translation = useAuthQuery(
     Queries.ai.translation({
       entry,
@@ -70,26 +50,6 @@ function EntryItemImpl({
     },
   )
 
-  const isActive = useRouteParamsSelector(
-    ({ entryId }) => entryId === entry.entries.id,
-  )
-
-  const asRead = useAsRead(entry)
-  const hoverMarkUnread = useGeneralSettingKey("hoverMarkUnread")
-
-  const handleMouseEnter = useDebounceCallback(
-    () => {
-      if (!hoverMarkUnread) return
-      if (!document.hasFocus()) return
-      if (asRead) return
-
-      entryActions.markRead(entry.feedId, entry.entries.id, true)
-    },
-    233,
-    {
-      leading: false,
-    },
-  )
   let Item: EntryListItemFC
 
   switch (view) {
@@ -122,82 +82,15 @@ function EntryItemImpl({
     }
   }
 
-  const navigate = useNavigateEntry()
-  const handleClick: React.MouseEventHandler<HTMLDivElement> = useCallback(
-    (e) => {
-      e.stopPropagation()
-
-      navigate({
-        entryId: entry.entries.id,
-      })
-    },
-    [entry.entries.id, navigate],
-  )
-  const handleDoubleClick: React.MouseEventHandler<HTMLDivElement> =
-    useCallback(
-      () => entry.entries.url && window.open(entry.entries.url, "_blank"),
-      [entry.entries.url],
-    )
-  const handleContextMenu: React.MouseEventHandler<HTMLDivElement> =
-    useCallback(
-      (e) => {
-        e.preventDefault()
-
-        showNativeMenu(
-          [
-            ...items
-              .filter((item) => !item.hide)
-              .map((item) => ({
-                type: "text" as const,
-                label: item.name,
-                click: item.onClick,
-                shortcut: item.shortcut,
-              })),
-            {
-              type: "separator" as const,
-            },
-            ...feedItems.filter((item) => !item.disabled),
-
-            {
-              type: "separator" as const,
-            },
-            {
-              type: "text" as const,
-              label: "Copy Entry ID",
-              click: () => {
-                navigator.clipboard.writeText(entry.entries.id)
-              },
-            },
-          ],
-          e,
-        )
-      },
-      [items, feedItems, entry.entries.id],
-    )
-
   return (
-    <div data-entry-id={entry.entries.id}>
-      <div
-        className={cn(
-          "relative rounded-md bg-theme-background transition-colors",
-          asRead ?
-            "text-zinc-700 dark:text-neutral-400" :
-            "text-zinc-900 dark:text-neutral-300",
-          Item.wrapperClassName,
-        )}
-        onClick={handleClick}
-        onMouseEnter={handleMouseEnter}
-        onDoubleClick={handleDoubleClick}
-        onContextMenu={handleContextMenu}
-      >
-        <ListItemHoverOverlay
-          className={cn(views[view || 0].gridMode ? "inset-0" : "inset-y-1")}
-          isActive={isActive}
-        >
-          <Item entryId={entry.entries.id} translation={translation.data} />
-        </ListItemHoverOverlay>
-      </div>
-    </div>
+    <EntryItemWrapper
+      itemClassName={Item.wrapperClassName}
+      entry={entry}
+      view={view}
+      overlay
+    >
+      <Item entryId={entry.entries.id} translation={translation.data} />
+    </EntryItemWrapper>
   )
 }
 
@@ -251,21 +144,3 @@ const createSkeletonItems = (element: ReactNode) => {
   }
   return children
 }
-
-export const EntryItemSkeletonWithDelayShow: FC<{
-  view: FeedViewType
-}> = memo(({ view }) => {
-  const SkeletonItem = SkeletonItemMap[view]
-
-  return SkeletonItem ?
-      (
-        <div
-          className={cn(views[view].gridMode ? girdClassNames : "flex flex-col")}
-        >
-          {createSkeletonItems(SkeletonItem)}
-        </div>
-      ) :
-      (
-        LoadingCircleFallback
-      )
-})
