--- src/renderer/src/modules/entry-column/readme.md
@@ -1,5 +1,0 @@
-Mark as read has three logic now.
-
-- Rendering entry as read.
-- Hover entry as read
-- Scrolling entry as read
