+++ src/renderer/src/pages/(external)/(with-layout)/feed/[id]/index.tsx
@@ -7,12 +7,12 @@ import { views } from "@renderer/constants"
 import { useTitle } from "@renderer/hooks/common"
 import { FeedViewType } from "@renderer/lib/enum"
 import { cn, pluralize } from "@renderer/lib/utils"
-import { ArticleItem } from "@renderer/modules/entry-column/article-item"
-import { NotificationItem } from "@renderer/modules/entry-column/notification-item"
-import { PictureItem } from "@renderer/modules/entry-column/picture-item"
-import { SocialMediaItem } from "@renderer/modules/entry-column/social-media-item"
+import { ArticleItem } from "@renderer/modules/entry-column/Items/article-item"
+import { NotificationItem } from "@renderer/modules/entry-column/Items/notification-item"
+import { PictureItem } from "@renderer/modules/entry-column/Items/picture-item"
+import { SocialMediaItem } from "@renderer/modules/entry-column/Items/social-media-item"
+import { VideoItem } from "@renderer/modules/entry-column/Items/video-item"
 import type { UniversalItemProps } from "@renderer/modules/entry-column/types"
-import { VideoItem } from "@renderer/modules/entry-column/video-item"
 import { useEntriesPreview } from "@renderer/queries/entries"
 import { useFeed } from "@renderer/queries/feed"
 import { DEEPLINK_SCHEME } from "@shared/constants"
